adapted to implemented front Technological operations

+++ GrainElevatorAPI/Controllers/TechnologicalOperationController.cs
@@ -29,6 +29,7 @@ public class TechnologicalOperationController : ControllerBase
 	}
 
 	[HttpPost]
+	[Authorize(Roles = "Admin,Technologist,Accountant")]
 	public async Task<ActionResult<TechnologicalOperationDto>> CreateTechnologicalOperation(TechnologicalOperationCreateRequest request)
 	{
 		if (!ModelState.IsValid) return BadRequest(ModelState);
@@ -36,9 +37,13 @@ public class TechnologicalOperationController : ControllerBase
 		try
 		{
 			var cancellationToken = GetCancellationToken();
-			var newTechnologicalOperation = _mapper.Map<TechnologicalOperation>(request);
+			
 			var createdById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
-
+			if (createdById <= 0)
+				return Unauthorized(new { message = "Користувач не авторизований." });
+			
+			var newTechnologicalOperation = _mapper.Map<TechnologicalOperation>(request);
+			
 			var createdTechnologicalOperation = await _technologicalOperationService.CreateTechnologicalOperationAsync(newTechnologicalOperation, createdById, cancellationToken);
 			return CreatedAtAction(nameof(GetTechnologicalOperationById), new { id = createdTechnologicalOperation.Id },
 				_mapper.Map<TechnologicalOperationDto>(createdTechnologicalOperation));
@@ -52,6 +57,7 @@ public class TechnologicalOperationController : ControllerBase
 
 
 	[HttpGet]
+	[Authorize(Roles = "Admin,Technologist,Accountant,CEO")]
 	public async Task<ActionResult<IEnumerable<TechnologicalOperationDto>>> GetTechnologicalOperations([FromQuery] int page = 1, [FromQuery] int size = 10)
 	{
 		try
@@ -89,14 +95,29 @@ public class TechnologicalOperationController : ControllerBase
 	}
 
 	[HttpGet("search")]
-	public async Task<ActionResult<IEnumerable<TechnologicalOperationDto>>> SearchTechnologicalOperations(string title)
+	[Authorize(Roles = "Admin,Technologist,Accountant,CEO")]
+	public async Task<ActionResult<IEnumerable<TechnologicalOperationDto>>> SearchTechnologicalOperations(
+		[FromQuery] string? title,
+		[FromQuery] string? createdByName,
+		[FromQuery] int page = 1,
+		[FromQuery] int size = 10,
+		[FromQuery] string? sortField = null,
+		[FromQuery] string? sortOrder = null)
 	{
 		try
 		{
 			var cancellationToken = GetCancellationToken();
-			var TechnologicalOperations = await _technologicalOperationService.SearchTechnologicalOperation(title, cancellationToken);
+			var (filteredTechnologicalOperations, totalCount) = await _technologicalOperationService.SearchTechnologicalOperation(
+				title,
+				createdByName,
+				page, 
+				size,
+				sortField, sortOrder,
+				cancellationToken);
 			
-			var TechnologicalOperationDtos = _mapper.Map<IEnumerable<TechnologicalOperationDto>>(TechnologicalOperations);
+			var TechnologicalOperationDtos = _mapper.Map<IEnumerable<TechnologicalOperationDto>>(filteredTechnologicalOperations);
+			Response.Headers.Append("X-Total-Count", totalCount.ToString());
+			
 			return Ok(TechnologicalOperationDtos);
 		}
 		catch (Exception ex)
@@ -108,6 +129,7 @@ public class TechnologicalOperationController : ControllerBase
 	
 	
 	[HttpPut("{id}")]
+	[Authorize(Roles = "Admin,Technologist,Accountant")]
 	public async Task<IActionResult> UpdateTechnologicalOperation(int id, TechnologicalOperationCreateRequest request)
 	{
 		if (!ModelState.IsValid) return BadRequest(ModelState);
@@ -115,11 +137,18 @@ public class TechnologicalOperationController : ControllerBase
 		try
 		{
 			var cancellationToken = GetCancellationToken();
+			
+			var modifiedById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
+			if (modifiedById <= 0)
+				return Unauthorized(new { message = "Користувач не авторизований." });
+			
+
 			var technologicalOperationDb = await _technologicalOperationService.GetTechnologicalOperationByIdAsync(id, cancellationToken);
-			if (technologicalOperationDb == null) return NotFound($"Технологічної операції з ID {id} не знайдено.");
+			if (technologicalOperationDb == null)
+				return NotFound($"Технологічної операції з ID {id} не знайдено.");
 			
 			technologicalOperationDb.Title = request.Title;
-			var modifiedById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
+			
 			var updatedTechnologicalOperation = await _technologicalOperationService.UpdateTechnologicalOperationAsync(technologicalOperationDb, modifiedById, cancellationToken);
 			
 			return Ok(_mapper.Map<TechnologicalOperationDto>(updatedTechnologicalOperation));
@@ -133,16 +162,21 @@ public class TechnologicalOperationController : ControllerBase
 
 	
 	[HttpPatch("{id}/soft-remove")]
-	//[Authorize(Roles = "admin, laboratory")]
+	[Authorize(Roles = "Admin,Technologist,Accountant")]
 	public async Task<IActionResult> SoftDeleteTechnologicalOperation(int id)
 	{
 		try
 		{
 			var cancellationToken = GetCancellationToken();
-			var technologicalOperationDb = await _technologicalOperationService.GetTechnologicalOperationByIdAsync(id, cancellationToken);
-			if (technologicalOperationDb == null) return NotFound($"Технологічної операції з ID {id} не знайдено.");
-
+			
 			var removedById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
+			if (removedById <= 0)
+				return Unauthorized(new { message = "Користувач не авторизований." });
+			
+			var technologicalOperationDb = await _technologicalOperationService.GetTechnologicalOperationByIdAsync(id, cancellationToken);
+			if (technologicalOperationDb == null) 
+				return NotFound($"Технологічної операції з ID {id} не знайдено.");
+			
 			var removedTechnologicalOperation = await _technologicalOperationService.SoftDeleteTechnologicalOperationAsync(technologicalOperationDb, removedById, cancellationToken);
 			
 			return Ok(_mapper.Map<TechnologicalOperationDto>(removedTechnologicalOperation));
@@ -156,16 +190,20 @@ public class TechnologicalOperationController : ControllerBase
 
 
 	[HttpPatch("{id}/restore")]
-	//[Authorize(Roles = "admin, laboratory")]
+	[Authorize(Roles = "Admin")]
 	public async Task<IActionResult> RestoreRemovedTechnologicalOperation(int id)
 	{
 		try
 		{
 			var cancellationToken = GetCancellationToken();
+			
+			var restoredById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
+			if (restoredById <= 0)
+				return Unauthorized(new { message = "Користувач не авторизований." });
+			
 			var technologicalOperationDb = await _technologicalOperationService.GetTechnologicalOperationByIdAsync(id, cancellationToken);
 			if (technologicalOperationDb == null) return NotFound($"Технологічної операції з ID {id} не знайдено.");
 			
-			var restoredById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
 			var restoredTechnologicalOperation = await _technologicalOperationService.RestoreRemovedTechnologicalOperationAsync(technologicalOperationDb, restoredById, cancellationToken);
 			
 			return Ok(_mapper.Map<TechnologicalOperationDto>(restoredTechnologicalOperation));
@@ -178,12 +216,17 @@ public class TechnologicalOperationController : ControllerBase
 	}
 	
 	[HttpDelete("{id}/hard-remove")]
-	[Authorize(Roles = "admin")]
+	[Authorize(Roles = "Admin")]
 	public async Task<IActionResult> DeleteTechnologicalOperation(int id)
 	{
 		try
 		{
 			var cancellationToken = GetCancellationToken();
+			
+			var removedById = HttpContext.Session.GetInt32("EmployeeId").GetValueOrDefault();
+			if (removedById <= 0)
+				return Unauthorized(new { message = "Користувач не авторизований." });
+			
 			var success = await _technologicalOperationService.DeleteTechnologicalOperationAsync(id, cancellationToken);
 			if (!success) return NotFound($"Технологічної операції з ID {id} не знайдено.");
